CS 251 ex2
The user is going to input data about 0 or more students, followed by the input “# 0 0”. Each student input consists of their netid, midterm score, and final score. Example input:

lshu3 90 92
gma4 82 91
rahsan3 79 81
# 0 0
Your job is to write a C program (not C++) that inputs this data into an array of structs, prints them out in reverse order (1 per line) with each student’s exam average, followed by the overall averages for the midterm and final. Here’s the correct output for the input above:

rahsan3: 79,81,80.000000
gma4: 82,91,86.500000
lshu3: 90,92,91.000000
Midterm: 83.666667
Final: 88.000000
If no students are input, i.e. the input is simply

# 0 0 
Then the output should be

Midterm: N/A
Final: N/A
Declare an array of size 100 to store the student array; you may assume the input will never exceed 100 students. Each element of the array is a struct, defined as follows:

typedef struct Student
{
  char Netid[16];
  int  Midterm;
  int  Final;
} Student;
Note that space for the student’s Netid is already declared; assume 16 chars is sufficient. Since the string-based input is a single word (netid), it is safe to use scanf for obtaining the value. However, since we are mixing strings and integers, it’s best to discard the rest of the input line before calling scanf again. Here’s a good way to input data about one student:

char netid[16];
int  mid, fnl;

scanf("%s %d %d", netid, &mid, &fnl);
skipRestOfInput(stdin);
The function skipRestOfInput() is provided. Recall that strings in C are arrays of chars, and that in order to store the student’s netid into the array, you’ll need to use strcpy() to copy the string-based input value into the corresponding Netid field of the array:

strcpy(students[i].Netid, netid);
Your input loop should stop when the netid "#" is input. Since valid UIC netids start with a letter, it's sufficient to check the first character of the input netid to see if it's a '#':

while (netid[0] != '#')
Finally, note that the averages should be computed as real numbers, and output using %f. For example:

printf("Midterm: %f\n", midtermAverage);